Admin remoto

+++ src/views/admin/product/Edit.vue
@@ -10,16 +10,10 @@
     <b-row>
         <b-colxx xxs="12">
             <b-card class="mb-4" title="Novo produto">
-                {{product.type}}
                 <b-row>
                     <b-colxx md="4" lg="3">
                         <b-form-group label="Nome do Produto" class="has-float-label mb-4">
                             <b-form-input v-model="product.name" type="text" placeholder="Nome do Produto" />
-                        </b-form-group>
-                    </b-colxx>
-                    <b-colxx md="4" lg="3">
-                        <b-form-group label="SKU" class="has-float-label mb-4">
-                            <b-form-input v-model="product.sku" type="text" placeholder="SKU" />
                         </b-form-group>
                     </b-colxx>
                     <b-colxx md="4" lg="3">
@@ -32,26 +26,21 @@
                             </fieldset>
                         </b-form-group>
                     </b-colxx>
-                    <b-colxx md="4" lg="3">
-                        <b-form-group>
-                            <select class="form-control" v-model="product.type">
-                                <option value="1">Tipo do produto</option>
-                                <option value="1">Óculos</option>
-                                <option value="2">Lente</option>
-                                <option value="4">Tratamento</option>
-                            </select>
-                        </b-form-group>
-                    </b-colxx>
                 </b-row>
             </b-card>
-            
+
             <b-card class="mb-4" title="Cor" v-if="product.type == 1">
                 <b-row>
                     <b-colxx md="4" lg="3">
-                        <b-form-group label="Cor" class="has-float-label mb-4">
+                        <b-form-group label="SKU" class="has-float-label mb-1">
+                            <b-form-input v-model="color.sku" type="text" placeholder="SKU" />
+                        </b-form-group>
+                        <b-alert variant="info" show  dismissible>O mesmo do bling!</b-alert>
+
+                        <b-form-group label="Cor" class="has-float-label my-4">
                             <b-form-input v-model="color.name" type="text" placeholder="Cor" />
                         </b-form-group>
-                        <vue-dropzone 
+                        <vue-dropzone
                             @vdropzone-complete="completeUpload"
                             ref="myVueDropzone" id="dropzone" :options="dropzoneOptions"></vue-dropzone>
                         <button class="btn btn-success w-100 mt-3" @click="addImage">Adicionar Cor</button>
@@ -71,7 +60,7 @@
                 </b-row>
             </b-card>
 
-            <b-card class="mb-4" title="Tamanho" v-if="product.type == 1">
+            <b-card class="mb-4" title="Tamanho">
                 <b-row>
                     <b-colxx lg="3">
                         <div class="alert alert-info"><i class="glyph-icon simple-icon-info"></i> Insira uma medida e tecle Enter</div>
@@ -124,7 +113,8 @@ export default {
             baseURL,
             color: {
                 name: '',
-                image: ''
+                image: '',
+                sku: ''
             },
             file: null
         }
@@ -166,14 +156,16 @@ export default {
         `
         },
         addImage() {
-            if(this.color.name && this.color.image){
+            if(this.color.name && this.color.image && this.color.sku){
                 this.product.colors.push({
                     name: this.color.name,
                     image: this.color.image,
+                    sku: this.color.sku,
                 });
 
                 this.color.name = '';
                 this.color.image = '';
+                this.color.sku = '';
                 this.$refs.myVueDropzone.removeFile(this.file);
             }else{
                 this.$notify("error", "Opsss...!", "Você precisa inserir uma imagem e o nome da cor", {
@@ -205,7 +197,7 @@ export default {
             const id = this.$route.params.id;
             const response = await api.get(`admin/product/${id}`);
             this.product = response.data.data;
-        } 
+        }
     },
     created(){
         this.getProduct();
